Fix duplicate additional ans and extra containers
When answering SRV questions, I noticed that duplicate A & AAAA records
were returned in the additional section of the answer. This commit fixes
the issue.

Since we are only interested in accessing containers that don't share an
IP address with the host, get processes that are in just unique net
namespaces, instead of the inique set of net, uts, and mount namespaces.